tile_service/requirements.txt
Pillow~=11.2.1
httpx~=0.28.1
redis~=6.2.0
rq~=2.3.3
python-dotenv~=1.1.0